requirements.txt
kafka-python==2.0.2
elasticsearch==7.13.0
pyspark==3.5.0
scikit-learn==1.3.2
